+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Implement collector for ps-constants project (OSIDB-1199)
 - Validate summary and requires_summary (OSIDB-1164)
+- Validate impact and summary (OSIDB-1164)
 
 ## [3.4.1] - 2023-08-21
 ### Changed